agents/qc_domain/requirements.txt
pandas>=2.1
tldextract>=5.1
requests>=2.31
beautifulsoup4>=4.12
rapidfuzz>=3.9
duckduckgo_search>=6.3.7
python-dotenv>=1.0
tqdm>=4.66
openpyxl>=3.1
XlsxWriter>=3.2
streamlit>=1.37
